+++ src/ScottPlot5/ScottPlot5 Demos/ScottPlot5 WinForms Demo/Demos/DataStreamer.cs
@@ -32,7 +32,7 @@ public partial class DataStreamer : Form, IDemoWindow
         {
             if (Streamer.HasNewData)
             {
-                formsPlot1.Plot.Title($"Processed {Streamer.DataSource.CountTotal:N0} points");
+                formsPlot1.Plot.Title($"Processed {Streamer.Data.CountTotal:N0} points");
                 formsPlot1.Refresh();
             }
         };