+++ src/ScottPlot5/ScottPlot5/DataViews/Scroll.cs
@@ -16,18 +16,18 @@ public class Scroll : IDataStreamerView
 
     public void Render(RenderPack rp)
     {
-        Pixel[] points = new Pixel[Streamer.DataSource.Length];
+        Pixel[] points = new Pixel[Streamer.Data.Length];
 
-        int oldPointCount = Streamer.DataSource.Length - Streamer.DataSource.NextIndex;
+        int oldPointCount = Streamer.Data.Length - Streamer.Data.NextIndex;
 
-        for (int i = 0; i < Streamer.DataSource.Length; i++)
+        for (int i = 0; i < Streamer.Data.Length; i++)
         {
             bool isNewPoint = i < oldPointCount;
-            int sourceIndex = isNewPoint ? Streamer.DataSource.NextIndex + i : i - oldPointCount;
-            int targetIndex = NewOnRight ? i : Streamer.DataSource.Data.Length - 1 - i;
+            int sourceIndex = isNewPoint ? Streamer.Data.NextIndex + i : i - oldPointCount;
+            int targetIndex = NewOnRight ? i : Streamer.Data.Data.Length - 1 - i;
             points[targetIndex] = new(
-                x: Streamer.Axes.GetPixelX(targetIndex * Streamer.DataSource.SamplePeriod + Streamer.DataSource.OffsetX),
-                y: Streamer.Axes.GetPixelY(Streamer.DataSource.Data[sourceIndex] + Streamer.DataSource.OffsetY));
+                x: Streamer.Axes.GetPixelX(targetIndex * Streamer.Data.SamplePeriod + Streamer.Data.OffsetX),
+                y: Streamer.Axes.GetPixelY(Streamer.Data.Data[sourceIndex] + Streamer.Data.OffsetY));
         }
 
         using SKPaint paint = new();